\name{anscombe}
\alias{anscombe}
\docType{data}
\title{Data created by Anscombe}
\description{
  A beautifully constructed data set for showing the need to look at data
}
\usage{data(anscombe)}
\format{
  A data frame with 44 observations on the following 3 variables.
  \describe{
    \item{\code{Food}}{a numeric vector}
    \item{\code{Drink}}{a numeric vector}
    \item{\code{Type}}{a factor with levels \code{bird} \code{insect} \code{mammal} \code{reptile}}
  }
}
\references{
  Anscombe, F.J. (1973). Graphs in statistical analysis. 
  \emph{American Statistician, 27}, 17-21.
}
\keyword{datasets}
