% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgpdump-package.R
\docType{package}
\name{bgpdump}
\alias{bgpdump}
\alias{bgpdump-package}
\title{Tools to work with 'BGP' 'dump' files}
\description{
A compendium of utilities to retrieve and process BGP files. Presently supports
download and converting MRT v2 files from RouteViews.
}
\author{
Bob Rudis (@hrbrmstr)
}

