\name{chile}
\alias{chile}
\docType{data}
\title{Chile length and Heat}
\description{
  The smaller the chile the hotter the taste. These data were gather from
  "Pepper Joe's" web page several years ago (\url{http://www.pepperjoe.com/}).
  This is a \emph{cool} example for teaching (imho).
}
\usage{data(chile)}
\format{
  A data frame with 85 observations on the following 3 variables.
  \describe{
    \item{\code{NAME}}{a factor with levels \code{Afric} \code{Aji} \code{Aji_A} \code{Aji_C} \code{Aji_F} \code{Aji_L} \code{Aji_M} \code{Aji_N} \code{Aji_O} \code{Aji_P} \code{Aji_R} \code{Aji_V} \code{Akola} \code{Amari} \code{Anahe} \code{Ancho} \code{Assam} \code{Azr} \code{Banan} \code{Bedgi} \code{Bell_} \code{Berbe} \code{Big_J} \code{Bird} \code{Bisho} \code{Black} \code{Boliv} \code{Bomba} \code{Bulga} \code{Cabai} \code{Calif} \code{Calor} \code{Canti} \code{Carib} \code{Casab} \code{Casca} \code{Cayen} \code{Charl} \code{Cherr} \code{Chile} \code{Chipo} \code{Cisco} \code{Confe} \code{de_Ar} \code{Dundi} \code{Dutch} \code{Fajit} \code{Fatal} \code{Garde} \code{Georg} \code{Goat_} \code{Guaji} \code{Haban} \code{Hunga} \code{Inca_} \code{Infer} \code{Jalap} \code{Kahar} \code{Kashm} \code{Kenya} \code{Large} \code{Long_} \code{Louis} \code{Malib} \code{NuMex} \code{Pasil} \code{Patzc} \code{Penis} \code{Rocot} \code{Sandi} \code{Sappo} \code{Scotc} \code{Serra} \code{Tabas}}
    \item{\code{LENGTH}}{a numeric vector}
    \item{\code{HEAT}}{a numeric vector}
  }
}
\keyword{datasets}
