\name{cp85}
\alias{cp85}
\docType{data}
\title{Wages for males and females}
\description{
  Data of wages with education, gender, and experience
}
\usage{data(cp85)}
\format{
  A data frame with 533 observations on the following 6 variables.
  \describe{
    \item{\code{EDUC}}{a numeric vector}
    \item{\code{FEMALE}}{a numeric vector}
    \item{\code{EXPER}}{a numeric vector}
    \item{\code{WAGE}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{LNWAGE}}{a numeric vector}
  }
}
\references{
  http://lib.stat.cmu.edu/datasets/CPS_85_Wages }
\keyword{datasets}
