\name{crime}
\alias{crime}
\docType{data}
\title{Crime in Sussex (UK)}
\description{
  These were taken from the Argos newspaper in Brighton. They show crime
  levels in different areas and in Wright and London (2009) are used to
  illustrate transformations, robust correlation, and robust regression.
}
\usage{data(crime)}
\format{
  A data frame with 261 observations on the following 10 variables.
  \describe{
    \item{\code{Neigh}}{a factor with levels \code{AldwickEast} \code{AldwickWest} \code{Alfriston} \code{Angmering} \code{Ardingly} \code{Arundel} \code{Ashdown} \code{AshurstWood} \code{Baird} \code{Barcombe} \code{Barnham} \code{BattleTown} \code{Beach(Littlehampton)} \code{Bersted} \code{Bewbush} \code{Billinghurst} \code{Bolney} \code{Bosham} \code{Boxgrove} \code{Bramber} \code{Braybrooke} \code{BredeValley} \code{BroadbrigeHeath} \code{BroadfieldNorth} \code{BroadfieldSouth} \code{BroadwaterDistrict} \code{Brookfield} \code{BrunswickandAdelaide} \code{Buckingham} \code{BurgessHillDunstall} \code{BurgessHIllFranklands} \code{BurgessHillLeylands} \code{BurgessHillMeeds} \code{BurgessHillStAndrews} \code{BurgessHillVictoria} \code{Bury} \code{Buxted} \code{Castle} \code{CastleDistrict} \code{Central} \code{CentralDistrict} \code{CentralHove} \code{CentralStLeonards} \code{Chailey} \code{Chanctonbury} \code{Chantry} \code{ChichesterEast} \code{ChichesterNorth} \code{ChichesterSouth} \code{ChichesterWest} \code{Chiddingly} \code{Churchill} \code{Cokeham} \code{Collington} \code{Conquest} \code{Copthorne} \code{Cowfod} \code{CrawledyDown} \code{CrossinHand} \code{CrowboroughEast} \code{CrowboroughJarvisBrook} \code{CrowboroughNorth} \code{CrowboroughStJohns} \code{CrowboroughWest} \code{Crowhurst} \code{Cuckfield} \code{Darweel} \code{Devonshire} \code{Ditchling} \code{Donnington} \code{DurringtonDistrict} \code{Earnley} \code{Easebourne} \code{EastBrighton} \code{Eastbrook} \code{EasternRother} \code{EastGrinsteadAshplats} \code{EastGrinsteadBaldwins} \code{EastGrinsteadHerontye} \code{EastGrinsteadImberhorn} \code{EastGrinsteadTown} \code{EastPreston} \code{EastSaltdean} \code{EastWittering} \code{EwhurstandSedlescombe} \code{FelphamEast} \code{FelphamWest} \code{Fernhurst} \code{Ferring} \code{Findon} \code{Fishbourne} \code{ForestRow} \code{Framfield} \code{Frant} \code{Funtington} \code{FurnaceGreen} \code{GaisfordDistrict} \code{Gensing} \code{Goldsmid} \code{GoringDistrict} \code{GosspopsGreen} \code{HailshamCentral} \code{HailshamEast} \code{HailshamSouth} \code{Ham} \code{HampdenPark} \code{Hampr} \code{HangletonandKnoll} \code{HanoverandElmGrove} \code{Hartfield} \code{Harting} \code{Hassocks} \code{HaywardsHeathAshengrou} \code{HaywardsHeathBentswood} \code{HaywardsHeathFrankland} \code{HaywardsHeathHeath} \code{HaywardsHeathLucastes} \code{HeathfieldEast} \code{HeathfieldNorth} \code{HeeneDistrict} \code{Hellingly} \code{Henfield} \code{Herstmonceux} \code{HighWeald} \code{Hillside} \code{HollingburyandStanmer} \code{Hollington} \code{Horam} \code{HorshamDenne} \code{HorshamForest} \code{HorshamHolbrookEast} \code{HorshamHolbrookWest} \code{HorshamRiverside} \code{HorshamRoffeyNorth} \code{HorshamRoffeySouth} \code{HorshamTrafalger} \code{Hotham(Bognor)} \code{Hurstpierpoint} \code{Ifield} \code{Itchingfield} \code{Kewhurst} \code{Kingston} \code{LangleyGreen} \code{Langney} \code{Lavant} \code{LewesBridge} \code{LewesCastle} \code{LewesPriory} \code{Lindfield} \code{Maidenbower} \code{MarineDistrict} \code{Marsham} \code{MashBarn} \code{Mayfield} \code{MazeHill} \code{Meads} \code{MiddletononSea} \code{Midhurst} \code{MousecoombandBevendean} \code{NewhavenDenton} \code{NewhavenValley} \code{Newick} \code{NinfieldandHooe} \code{Northbrook} \code{Northgate} \code{NorthMundham} \code{NorthPortslade} \code{Nuthurst} \code{Offington} \code{OldHastings} \code{OldTown} \code{Orchard} \code{Ore} \code{OuseValley} \code{PaghamandRoseGreen} \code{Patcham} \code{PeacehavenEast} \code{PeacehavenNorth} \code{PeacehavenWest} \code{Petworth} \code{Pevensey} \code{PevenseyandWestham} \code{Peverel} \code{Plaistow} \code{Plumpton} \code{PolegateNorth} \code{PolegateSouth} \code{PoundHillNorth} \code{PoundHillSouth} \code{PrestonPark} \code{Pulborough} \code{QueensPark} \code{Ratton} \code{Regency} \code{River(Littlehampton)} \code{Rogate} \code{Rotherfield} \code{RotherLevels} \code{RottingdeanCoastal} \code{Rudgwisk} \code{RusperandColgate} \code{RustingtonEast} \code{RustingtonWest} \code{Rye} \code{Sackville} \code{Salehurst} \code{SalvingtonDistrict} \code{SeafordCentral} \code{SeafordEast} \code{SeafordNorth} \code{SeafordSouth} \code{SeafordWest} \code{Selden} \code{SelseyNorth} \code{SelseySouth} \code{Sidlesham} \code{Sidley} \code{Silverhill} \code{Southbourne} \code{Southgate} \code{Southlands} \code{SouthPortslade} \code{Southwate} \code{SouthwickGreen} \code{Sovereign} \code{Stanford} \code{StAnthonys} \code{Stedham} \code{Steyning} \code{StHelens} \code{StMarks} \code{StMarys} \code{StMichaels} \code{StNicholas} \code{StPetersandNorthLai} \code{StStephens} \code{Tangmere} \code{Tarrying} \code{ThreeBridges} \code{Ticehurst} \code{Tilgate} \code{Tressell} \code{UckfieldCentral} \code{UckfieldNewtown} \code{UckfieldNorth} \code{UckfieldRidgewood} \code{Upperton} \code{Wadhurst} \code{Westbourne} \code{WestGreen} \code{WestStLeonards} \code{WestWittering} \code{WickwithToddington} \code{Widewater} \code{Willingdon} \code{Wish} \code{WishboroughGreen} \code{WishingTree} \code{Withdean} \code{Woodingdean} \code{Yapton}}
    \item{\code{Council}}{a numeric vector}
    \item{\code{Weapons}}{a numeric vector}
    \item{\code{Robbery}}{a numeric vector}
    \item{\code{Burglary}}{a numeric vector}
    \item{\code{Drugs}}{a numeric vector}
    \item{\code{Cars}}{a numeric vector}
    \item{\code{Sexoff}}{a numeric vector}
    \item{\code{theft}}{a numeric vector}
    \item{\code{damage}}{a numeric vector}
  }
}
\references{
  The Argos, March 10, 2007
}
\keyword{datasets}
