\name{exer}
\alias{exer}
\docType{data}
\title{Getting Children to Exercise}
\description{
  These data are from Hill et al. (2007) and are used to illustrate multilevel modeling.
}
\usage{data(exer)}
\format{
  A data frame with 503 observations on the following 12 variables.
  \describe{
    \item{\code{wcond}}{a numeric vector}
    \item{\code{class}}{a numeric vector}
    \item{\code{w1sn02}}{a numeric vector}
    \item{\code{w2sn02}}{a numeric vector}
    \item{\code{w1int}}{a numeric vector}
    \item{\code{w2int}}{a numeric vector}
    \item{\code{w1att}}{a numeric vector}
    \item{\code{w2att}}{a numeric vector}
    \item{\code{z1pbc}}{a numeric vector}
    \item{\code{z2pbc}}{a numeric vector}
    \item{\code{sqw1}}{a numeric vector}
    \item{\code{sqw2}}{a numeric vector}
  }
}
\references{
	Hill, C., Abraham, C., & Wright, D. B. (2007). Can theory-based messages
  in combination with cognitive prompts promote exercise in classroom settings?
  \emph{Social Science & Medicine}, \bold{65}, 1049-1058.
}
\keyword{datasets}
