% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routeviews.r
\name{get_rib}
\alias{get_rib}
\title{Retrieve a specific, historical RouteViews RIB file}
\usage{
get_rib(date, hour, output_dir = ".", .progress = FALSE)
}
\arguments{
\item{date}{the date (`YYYY-mm-dd` string or \code{Date}) of the file (remembering
that the filenames are in UTC time zone)}

\item{hour}{the hour of the file (remembering that RouteViews dumps a RIB every 2 hours)}

\item{.progress}{show download progress (default: \code{FALSE})}
}
\value{
the filename (invisibly)
}
\description{
Retrieve a specific, historical RouteViews RIB file
}

