\name{glmexample}
\alias{glmexample}
\docType{data}
\title{Made-up data to illustrate GLMs}
\description{
  These data were first created for Wright (2006).
}
\usage{data(glmexample)}
\format{
  A data frame with 20 observations on the following 6 variables.
  \describe{
    \item{\code{test}}{a numeric vector}
    \item{\code{social}}{a numeric vector}
    \item{\code{books}}{a numeric vector}
    \item{\code{math}}{a numeric vector}
    \item{\code{pass}}{a numeric vector}
    \item{\code{detent}}{a numeric vector}
  }
}
\references{
 	Wright, D. B. (2006a). The art of statistics: A survey of modern statistics.
   In P. A. Alexander & P. H. Winne (Eds.),
   \emph{Handbook of educational psychology (2nd ed.)} (pp. 879-901). Mahwah, NJ: Erlbaum.
}
\keyword{datasets}
