\name{lordex}
\alias{lordex}
\docType{data}
\title{Illustrating Lord's Paradox}
\description{
  These are data based on London, Bruck and Melnyk (2008) on children's
  suggestibility and are used to illustrate Lord's paradox (Wright, 2006).
}
\usage{data(lordex)}
\format{
  A data frame with 51 observations on the following 3 variables.
  \describe{
    \item{\code{AGEMOS}}{a numeric vector}
    \item{\code{Final}}{a numeric vector}
    \item{\code{Initial}}{a numeric vector}
  }
}
\references{
	London, K., Bruck, M., & Melnyk, L. (2008). Persistence of facilitation
  and misinformation effects in event memory following a 10 month delay.
  \emph{Law and Human Behavior}.
  	Wright, D. B. (2006b). Comparing groups in a before-after design:
    When t-test and ANCOVA produce different results.
    \emph{British Journal of Educational Psychology}, \bold{76}, 663-675.
}
\keyword{datasets}
