\name{maleptsd}
\alias{maleptsd}
\docType{data}
\title{Male PTSD symptoms after childbirth}
\description{
  These data come from ongoing research by Susan Ayers and colleagues on
  psychological well-being after
  childbirth (\url{http://www.birthtraumaassociation.org.uk/index.htm}). These
  data were used to illustrate model selection and regularization.}
\usage{data(maleptsd)}
\format{
  A data frame with 64 observations on the following 11 variables.
  \describe{
    \item{\code{OVER2}}{a numeric vector}
    \item{\code{OVER3}}{a numeric vector}
    \item{\code{OVER5}}{a numeric vector}
    \item{\code{BOND}}{a numeric vector}
    \item{\code{POSIT}}{a numeric vector}
    \item{\code{NEG}}{a numeric vector}
    \item{\code{CONTR}}{a numeric vector}
    \item{\code{SUP}}{a numeric vector}
    \item{\code{CONS}}{a numeric vector}
    \item{\code{AFF}}{a numeric vector}
    \item{\code{PTSD}}{a numeric vector}
  }
}
\references{
Ayers, S., Wright, D. B., & Wells N. (2007). Post-traumatic stress in
couples after birth: Association with the couple's relationship and parent-baby bond.
\emph{Journal of Reproductive & Infant Psychology}, \bold{25}, 40-50.
}
\keyword{datasets}
