\name{mrt-package}
\alias{mrt-package}
\alias{mrt}
\docType{package}
\title{
Modern Regression Techniques: Datasets and miscellaneous
}
\description{
Datasets and assorted functions for Wright and London's (2009)
\emph{Modern Regression Techniques Using R} and other materials. Other functions
are described in Kelley & Wright (2010), Wright (2007a,b), and Wright & Herrinton
(2011). The functions for Wright, Horry and Skagerberg (2009) are in sdtalt.
}
\details{
\tabular{ll}{
Package: \tab modreg\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2010-11-24\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel B. Wright <dwright@fiu.edu>
}
\references{
Kelley, M. R. & Wright, D.B. (2010) Obtaining representative nominal groups. 
\emph{Behavior Research Methods}, \bold{42}, 36-41.
\cr Wright, D.B. (2007a). Calculating nominal group statistics in collaboration studies.
\emph{Behavior Research Methods}, \bold{39}, 460-470.
\cr Wright, D. B. (2007b). Graphing within subject confidence intervals (WSCI) 
using SPSS and S-Plus. \emph{Behavior Research Methods}, \bold{39}, 82-85.
\cr Wright, D.B. & Harrington, J.A. (2011, actual in press at the moment). 
Problematic standard errors and confidence intervals for skewness and 
kurtosis. \emph{Behavior Research Methods}. www2.fiu.edu/~dwright/skewkurt
\cr Wright, D.B. & London, K. (2009). \emph{Modern regression techniques using R}. Sage Publications.
}
\keyword{ package }
