% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rib_to_asn_table}
\alias{rib_to_asn_table}
\title{Convert RouteViews RIB (bgpdump table dump v2) to subnet/asn map}
\usage{
rib_to_asn_table(path, progress = FALSE)
}
\arguments{
\item{path}{path to RouteViews RIB (can be compressed)}

\item{progress}{if \code{TRUE} output a message every 50,000 records.}
}
\value{
\code{tbl_df} (tibble) or \code{NULL}
}
\description{
It presently ONLY works with the modern RouteViews v2 format.
}
