\name{vconf}
\alias{vconf}
\title{Non-bootstrap Cramer's V intervals using non-central chi-square}
\description{Follows Smithson producing the confidence intervals for
Cramers V using noncentral chi-square. A boot function for Cramers V 
will arrive at some point}
\usage{vconf(ctab, clevel=.95)}
\arguments{
  \item{ctab}{the rxc contingency table}
  \item{clevel}{Confidence level}
}
\references{Smithson}