\name{mrt-package}
\alias{mrt-package}
\alias{mrt}
\docType{package}
\title{
Modern Regression Techniques: Datasets
}
\description{
Datasets and assorted functions for Wright and London's (2009)
\emph{Modern Regression Techniques Using R}.
}
\details{
\tabular{ll}{
Package: \tab modreg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-09\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel B. Wright <dwright@fiu.edu>
}
\references{
Wright, D.B. & London, K. (2009). \emph{Modern regression techniques using R}. Sage Publications.
}
\keyword{ package }
